+++ tests/Bowl/BowlClosureTest.php
@@ -58,9 +58,6 @@ class BowlClosureTest extends AbstractBowlTests
         return ['bar' => 'foo', 'bar2' => ['bar', 'foo']];
     }
 
-    /**
-     * @inheritDoc
-     */
     public function buildBowl(): BowlInterface
     {
         return new Bowl(