+++ tests/Bowl/BowlObjectTest.php
@@ -63,9 +63,6 @@ class BowlObjectTest extends AbstractBowlTests
         return ['bar' => 'foo', 'bar2' => ['bar', 'foo']];
     }
 
-    /**
-     * @inheritDoc
-     */
     public function buildBowl(): BowlInterface
     {
         return new Bowl(