+++ tests/Bowl/BowlParameterSelectedOnTypeTest.php
@@ -69,9 +69,6 @@ class BowlParameterSelectedOnTypeTest extends AbstractBowlTests
         return ['bar' => 'foo', 'bar2' => ['bar', 'foo']];
     }
 
-    /**
-     * @inheritDoc
-     */
     public function buildBowl(): BowlInterface
     {
         return new Bowl(