+++ tests/Bowl/BowlWithUnionTest.php
@@ -68,9 +68,6 @@ EOF;
         return ['bar' => 'foo', 'bar2' => ['bar', 'foo']];
     }
 
-    /**
-     * @inheritDoc
-     */
     public function buildBowl(): BowlInterface
     {
         return new Bowl(